+++ src/__tests__/discriminatedUnions.test.ts
@@ -1,31 +1,31 @@
 import { makeModel, ModelInstance } from '../model'
 
-export interface CarTraits {
+interface CarTraits {
   tag: 'Car'
   make: string
   model: string
 }
 
-export interface Car extends ModelInstance<CarTraits> {}
+interface Car extends ModelInstance<CarTraits> {}
 
-export interface BoatTraits {
+interface BoatTraits {
   tag: 'Boat'
   name: string
 }
 
-export interface Boat extends ModelInstance<BoatTraits> {}
+interface Boat extends ModelInstance<BoatTraits> {}
 
-export interface PlaneTraits {
+interface PlaneTraits {
   tag: 'Plane'
 }
 
-export interface Plane extends ModelInstance<PlaneTraits> {}
+interface Plane extends ModelInstance<PlaneTraits> {}
 
-export type VehicleTraits = CarTraits | BoatTraits | PlaneTraits
+type VehicleTraits = CarTraits | BoatTraits | PlaneTraits
 
-export type Vehicle = ModelInstance<VehicleTraits>
+type Vehicle = ModelInstance<VehicleTraits>
 
-export const makeVehicle = makeModel<VehicleTraits>()
+const makeVehicle = makeModel<VehicleTraits>()
 
 describe('Discriminated Unions', () => {
   it('correctly infers the type of `with` for discriminated unions', () => {