+++ src/model.ts
@@ -1,9 +1,5 @@
-import produce, { setAutoFreeze } from 'immer'
-import cloneDeep from 'lodash.clonedeep'
+import produce from 'immer'
 import { Lens, Lensed } from './lens'
-
-// Disable auto-freezing so that we can use `with` and `withMutations` interchangeably.
-setAutoFreeze(false)
 
 export type Transform<T> = (model: T) => void
 
@@ -14,18 +10,6 @@ export interface Model<T> {
    * @param transform The transform to apply to the model.
    */
   with(transform: Transform<T>): T & Model<T>
-
-  /**
-   * Runs the given transform and mutates the model in-place.
-   *
-   * @param transform The transform to apply to the model.
-   */
-  withMutations(transform: Transform<T>): T & Model<T>
-
-  /**
-   * Returns a clone of the model.
-   */
-  clone(): T & Model<T>
 
   /**
    * Views the model through the given lens.
@@ -41,13 +25,6 @@ export const Model = <T>(props: T): ModelInstance<T> => ({
   ...(props as any),
   with(transform: Transform<T>) {
     return produce(transform)(this as any)
-  },
-  withMutations(transform: Transform<T>) {
-    transform(this as any)
-    return this
-  },
-  clone() {
-    return cloneDeep(this)
   },
   throughLens<TLens>(lens: Lens<T, TLens>) {
     return lens(this)