+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Added ability to specify default properties through `makeModel`
 
+### Removed
+
+* `Model.withMutations`
+* `Model.clone`
+* `Lens.withMutations`
+
 ## [0.3.2] - 2018-04-03
 
 ### Fixed